Move referenceholder into the utils
